backend/requirements.txt
# Backend dependencies — Sistema de Reservas Guardería Canina
# Instalar con: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos / Supabase
supabase==2.4.3

# Variables de entorno
python-dotenv==1.0.1

# Validación de datos
pydantic==2.7.1

# Formularios multipart (subida de archivos)
python-multipart==0.0.9

# HTTP client para tests de integración
httpx==0.27.0

# Exportación
openpyxl==3.1.2
reportlab==4.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
hypothesis==6.100.1
